+++ components/PageHeader.vue
@@ -36,7 +36,7 @@ export default class PageHeader extends Vue {
   height: $header-height
   background-color: transparentize($sch-gray4, 0.1)
   color: $white
-  z-index: 999
+  z-index: 1000
   display: flex
   justify-content: space-between
   align-items: center