+++ pages/index.vue
@@ -9,7 +9,8 @@
         :class="{ 'nuxt-link-exact-active': linkObject.customExactActive }"
       ) {{ year }}
   .container-fluid.color-purple
-    .scholars-list(v-if="currentScholars.length > 0")
+    .loading-scholars(v-if="$fetchState.pending && currentScholars.length < 1") #[i Gathering the brightest minds from around the world]&nbsp;&nbsp;🤓
+    .scholars-list(v-else-if="currentScholars.length > 0")
       scholar-thumbnail(
         v-for="scholar in currentScholars",
         :scholar.once="scholar",
@@ -77,9 +78,11 @@ export default class PageIndex extends Vue {
   get currentScholars(): Scholar[] {
     if (!this.currentYear) return []
     const currentYearRecordName = this.currentYear.recordName
-    return Object.values(this.allScholars).filter(scholar => {
-      return scholar.wwdcYears && scholar.wwdcYears.some(ref => ref.recordName === currentYearRecordName)
-    })
+    return Object.values(this.allScholars)
+      .filter(scholar => {
+        return scholar.wwdcYearsApproved && scholar.wwdcYearsApproved.some(yearReference => yearReference.recordName === currentYearRecordName)
+      })
+      .sort((lhs, rhs) => lhs.givenName.localeCompare(rhs.givenName))
   }
 
   validate({ params }): boolean {
@@ -88,25 +91,27 @@ export default class PageIndex extends Vue {
     return /^\d{4}$/.test(year)
   }
 
-  async fetch ({ store, params, error }) {
-    await store.dispatch('years/queryYears')
-    const years = store.state.years.years
+  async fetch() {
+    await this.$store.dispatch('years/queryYears')
+    const years = this.$store.state.years.years
 
     let year: WWDCYear
-    if (params.year) {
-      year = years[`WWDC ${params.year}`]
+
+    if (this.$route.params.year) {
+      year = years[`WWDC ${this.$route.params.year}`]
     } else {
-      const keys = store.getters['years/sortedKeys']
+      const keys = this.$store.getters['years/sortedKeys']
       year = years[keys[keys.length - 1]]
     }
     if (!year) {
-      return error({
-        statusCode: 404,
-        message: 'The requested year could not be found in our database.'
-      })
+      if (process.server) {
+        this.$nuxt.context.res.statusCode = 404
+      }
+
+      throw new Error('The requested year could not be found in our database.')
     }
 
-    await store.dispatch('scholars/queryScholars', year)
+    await this.$store.dispatch('scholars/queryScholars', year)
   }
 }
 </script>
@@ -123,7 +128,7 @@ export default class PageIndex extends Vue {
   .scholar-thumbnail
     display: block
 
-.no-scholars
+.no-scholars, .loading-scholars
   margin-top: 30px
   text-align: center
   color: $sch-gray0